+++ backend/src/controllers/countryController.ts
@@ -6,6 +6,7 @@ import * as helper from '../utils/helper'
 import * as env from '../config/env.config'
 import i18n from '../lang/i18n'
 import Country from '../models/Country'
+import User from '../models/User'
 import LocationValue from '../models/LocationValue'
 import Location from '../models/Location'
 import * as logger from '../utils/logger'
@@ -61,8 +62,8 @@ export const validate = async (req: Request, res: Response) => {
       res.sendStatus(200)
     }
   } catch (err) {
-    logger.error(`[country.validate]  ${i18n.t('DB_ERROR')} ${name}`, err)
-    res.status(400).send(i18n.t('DB_ERROR') + err)
+    logger.error(`[country.validate]  ${i18n.t('ERROR')} ${name}`, err)
+    res.status(400).send(i18n.t('ERROR') + err)
   }
 }
 
@@ -94,8 +95,8 @@ export const create = async (req: Request, res: Response) => {
     await country.save()
     res.send(country)
   } catch (err) {
-    logger.error(`[country.create] ${i18n.t('DB_ERROR')} ${JSON.stringify(req.body)}`, err)
-    res.status(400).send(i18n.t('DB_ERROR') + err)
+    logger.error(`[country.create] ${i18n.t('ERROR')} ${JSON.stringify(req.body)}`, err)
+    res.status(400).send(i18n.t('ERROR') + err)
   }
 }
 
@@ -115,6 +116,16 @@ export const update = async (req: Request, res: Response) => {
     const country = await Country.findById(id).populate<{ values: env.LocationValue[] }>('values')
 
     if (country) {
+      // begin of security check
+      const sessionUserId = req.user?._id
+      const sessionUser = await User.findById(sessionUserId)
+      if (!sessionUser || (sessionUser.type === bookcarsTypes.UserType.Supplier && country.supplier?.toString() !== sessionUserId)) {
+        logger.error(`[country.update] Unauthorized attempt to update country ${country._id} by user ${sessionUserId}`)
+        res.status(403).send('Forbidden: You cannot update this country')
+        return
+      }
+      // end of security check
+
       const names: bookcarsTypes.CountryName[] = req.body
 
       for (const name of names) {
@@ -139,8 +150,8 @@ export const update = async (req: Request, res: Response) => {
     logger.error('[country.update] Country not found:', id)
     res.sendStatus(204)
   } catch (err) {
-    logger.error(`[country.update] ${i18n.t('DB_ERROR')} ${JSON.stringify(req.body)}`, err)
-    res.status(400).send(i18n.t('DB_ERROR') + err)
+    logger.error(`[country.update] ${i18n.t('ERROR')} ${JSON.stringify(req.body)}`, err)
+    res.status(400).send(i18n.t('ERROR') + err)
   }
 }
 
@@ -164,12 +175,22 @@ export const deleteCountry = async (req: Request, res: Response) => {
       res.status(204).send(msg)
       return
     }
+    // begin of security check
+    const sessionUserId = req.user?._id
+    const sessionUser = await User.findById(sessionUserId)
+    if (!sessionUser || (sessionUser.type === bookcarsTypes.UserType.Supplier && country.supplier?.toString() !== sessionUserId)) {
+      logger.error(`[country.delete] Unauthorized attempt to delete country ${country._id} by user ${sessionUserId}`)
+      res.status(403).send('Forbidden: You cannot delete this country')
+      return
+    }
+    // end of security check
+
     await Country.deleteOne({ _id: id })
     await LocationValue.deleteMany({ _id: { $in: country.values } })
     res.sendStatus(200)
   } catch (err) {
-    logger.error(`[country.delete] ${i18n.t('DB_ERROR')} ${id}`, err)
-    res.status(400).send(i18n.t('DB_ERROR') + err)
+    logger.error(`[country.delete] ${i18n.t('ERROR')} ${id}`, err)
+    res.status(400).send(i18n.t('ERROR') + err)
   }
 }
 
@@ -205,8 +226,8 @@ export const getCountry = async (req: Request, res: Response) => {
     logger.error('[country.getCountry] Country not found:', id)
     res.sendStatus(204)
   } catch (err) {
-    logger.error(`[country.getCountry] ${i18n.t('DB_ERROR')} ${id}`, err)
-    res.status(400).send(i18n.t('DB_ERROR') + err)
+    logger.error(`[country.getCountry] ${i18n.t('ERROR')} ${id}`, err)
+    res.status(400).send(i18n.t('ERROR') + err)
   }
 }
 
@@ -289,8 +310,8 @@ export const getCountries = async (req: Request, res: Response) => {
 
     res.json(countries)
   } catch (err) {
-    logger.error(`[country.getCountries] ${i18n.t('DB_ERROR')} ${req.query.s}`, err)
-    res.status(400).send(i18n.t('DB_ERROR') + err)
+    logger.error(`[country.getCountries] ${i18n.t('ERROR')} ${req.query.s}`, err)
+    res.status(400).send(i18n.t('ERROR') + err)
   }
 }
 
@@ -396,8 +417,8 @@ export const getCountriesWithLocations = async (req: Request, res: Response) => 
 
     res.json(countries)
   } catch (err) {
-    logger.error(`[country.getCountries] ${i18n.t('DB_ERROR')} ${req.query.s}`, err)
-    res.status(400).send(i18n.t('DB_ERROR') + err)
+    logger.error(`[country.getCountries] ${i18n.t('ERROR')} ${req.query.s}`, err)
+    res.status(400).send(i18n.t('ERROR') + err)
   }
 }
 
@@ -428,8 +449,8 @@ export const checkCountry = async (req: Request, res: Response) => {
 
     res.sendStatus(204)
   } catch (err) {
-    logger.error(`[country.checkCountry] ${i18n.t('DB_ERROR')} ${id}`, err)
-    res.status(400).send(i18n.t('DB_ERROR') + err)
+    logger.error(`[country.checkCountry] ${i18n.t('ERROR')} ${id}`, err)
+    res.status(400).send(i18n.t('ERROR') + err)
   }
 }
 
@@ -456,7 +477,7 @@ export const getCountryId = async (req: Request, res: Response) => {
     }
     res.sendStatus(204)
   } catch (err) {
-    logger.error(`[country.getCountryId] ${i18n.t('DB_ERROR')} ${name}`, err)
-    res.status(400).send(i18n.t('DB_ERROR') + err)
+    logger.error(`[country.getCountryId] ${i18n.t('ERROR')} ${name}`, err)
+    res.status(400).send(i18n.t('ERROR') + err)
   }
 }
